backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and Caching
redis==5.0.1
aioredis==2.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# API and Data Processing
requests==2.31.0
aiohttp==3.9.1
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2

# Sentiment Analysis
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2

# Logging and Monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.14

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
structlog==23.2.0 